+++ pluslib/Support/Facades/Rest.php
@@ -1,9 +1,0 @@
-<?php
-namespace pluslib\Support\Facades;
-
-use pluslib\HTTP\RestAPI;
-
-class Rest extends Facade
-{
-  protected $accessor = 'rest';
-}