+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release Notes for 0.x
 
+## [Unreleased](https://github.com/parsa-mostafaie/pluslib/compare/v0.0.04...master)
+
+- Remove `ABSPATH` and `RELPATH`
+- Remove not-needed class and helpers
+
 ## [v0.0.04](https://github.com/parsa-mostafaie/pluslib/compare/v0.0.03...v0.0.0.04)
 
 - `Select::first`: apply default value